Fixed weird wrapping issue.
For some reason Asciidoctor was splitting the word 'provider' not on a word boundary. This only happened when it was in a link. I fiddled with it for a bit but couldn't find anything in the internet or with my tweaking to fix it, so just moved it to a sep line.

Doesn't solve the root issue, but is a patch.