Updated to 4.0.1 of jekyll.
This removes the 'Using the last argument as keyword parameters is deprecated' errors.

It also upgraded rouge, and that causes slight differences to a few files that were using lexers which are upgraded. I examined a few of them by eye (the guide to user data security, the php laravel app) and didn't see any issues.