Removing this hook.
This hook limits our search results to just our docs directory.

That was fine initially, but now we have search on all our pages.

So we want to be able to search them.

See https://discourse.algolia.com/t/only-display-results-from-one-subdirectory-path/10159/2 for more on how this was implemented.

Removing so we will process all files for the search index.