Changed wording.
Daniel pointed out that creating a certificate isn't really the point. Whether the cert is self signed or created and not in the java keystore, it doesn't matter. What matters is that you are providing a certificate and stepping outside the bounds of the normal java keystore/CA chain.

Updated the doc to reflect that.