Added spacing.
If the copy button was overwritten, it wouldn't work.

I tried messing with the CSS, adding a z-index, comparing it to the docs copy button, but couldn't figure it out.

Adding in this div causes the content of the code block to move down one line, always ensuring that the copy button/icon is visible and clickable.

Definitely a hack, but I don't know how else to do this.