run this on both PR open and push.
The issue is if you only run it on push, it misses the first time you push up new images (that is what happened here: https://github.com/FusionAuth/fusionauth-site/pull/3375 ) because there is no PR open at that time, so the script fails. See this run for example: https://github.com/FusionAuth/fusionauth-site/actions/runs/11731437467

So we need to run this both on PR open (so it catches this case) and on push, so that when an image is updated in a PR or a new image is added, we catch it.

There's some worry about degradation of images due to repeated shrinking, but I haven't seen that play out in practice