updated to reflect that FQDN should include prefix for OIDC compliance.
https://openid.net/specs/openid-connect-discovery-1_0.html#ProviderMetadata says

the issuer is 'REQUIRED. URL using the https scheme with no query or fragment components that the OP asserts as its Issuer Identifier. If Issuer discovery is supported (see Section 2), this value MUST be identical to the issuer value returned by WebFinger. This also MUST be identical to the iss Claim value in ID Tokens issued from this Issuer. '